+++ documentation/changelog.md
@@ -8,7 +8,6 @@
 - Report should use getSubProp
 - Video gets out of sync with anim if switch to different tab
 - Shape "line" not working as expected.
-- Authoring tool var broken
 
 #### B
 - Shape "bcurve" offseta is not defined.
@@ -16,6 +15,7 @@
 - Figure out why some objects using SVG images are not interactive.
 - Var set in authoring tool doesn't change based on entry type.
 - boolean var type not supported
+- Not a function authorLibs.utils.deletedriver
 
 #### C
 - Need to use end time for video
@@ -33,6 +33,7 @@
 
 
 ## Changelog:
+- 2018-12-03 Fixed constraint type properties in authoring tool.
 - 2018-12-03 Fixed var properties in authoring tool.
 - 2018-12-03 Fixed modvar to work with strings.
 - 2018-12-03 Added check to only activate object on top when clicking.